Render DataFusion metrics in aggregatescan `EXPLAIN ANALYZE`.
